mock-trade-api/app/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
sqlalchemy>=1.4
psycopg2-binary>=2.9
pydantic>=1.10
python-dotenv>=1.0
PyJWT>=2.0
